Fix package import error: merge setup.py into setup/__init__.py
- Move all functions from kuasarr/storage/setup.py to kuasarr/storage/setup/__init__.py
- Delete the old setup.py to resolve package shadowing
- Fixes: cannot import name 'hostname_form_html' from 'kuasarr.storage.setup'
